Fix circuit state flickering on autosave
Fixes an error where during autosave (or rather any save, but it
was only visible during an autosave) would result in the circuit
flickering into a different circuit in the project. Also, a self
balancing circuit would switch into its other legal state without
any driving signal.

This happens because of a force-reset-nodes call in the generate-
circuit-data workflow. I am not sure why this call exists, my
guess is it's related to Subcircuits 1.0, but it is redundant
and only causes problems now.

This commit removes the force-reset-nodes call by removing the
update-everything flag from the update call in the generate-
save-data workflow.

Signed-off-by: Ruturaj Mohite <mohite.ruturaj15@gmail.com>