Make data save functions read only
Instead of calling removeConnections, we now simply ignore
cross-scope connections.

Remove redundant call to update

The save functions should be strictly read-only and should not
change the state of the simulator. This is especially important
given the existence of auto-save. We don't want simulator state
to change every 3 seconds when auto-save is called even when the
user did not change the state.

For most circuits this is not a problem, but for some circuits
where a section of the circuit has multiple legal values, the
circuit should not change to the other legal value without a
driving signal. But a force-reset update might cause such a
state change. Earlier such an update call was a part of the save
flow and was called with autosave(), this should not be a problem
after this commit.