Deferred Contentions (#4882)
* Deferred Contentions

Before this commit, contentions were checked at output nodes
and if a disagreeing value was found (only exception being
Tristates), the simulation would error out with a contention
error. This is not correct since some contentions might occur
temporarily and might resolved in the same play() cycle down the
resolution queue.

This commit maintains an [effectively] set of all contentions
that are found during resolution. Some contentions might get
resolved in the simulation and they are removed from the set.
At the end of the play cycle, if any contentions remain, then only
those are real contentions and are reported.

This commit also removes the special contentionPending check for
Tristate as this should be handled by the new contention system.

* Add back Tristate special check

Turns out we need it.