Do not check contention on downstream values
We define an output node value to be upstream if it is given to
it by its parent element. All other values are downstream values.

In the case that a node gives an output node value (while the
parent does not drive the output value. e.g. Tristate with control
bit 0), and then the same node gives another value to the output
node, earlier it would cause contention since the output node is
said to [before the propagation from the other node] still have
the former value.

This is a downstream value and post this commit will be
overwritten. Contention will only be checked at upstream valued
output nodes.