Fix: SubCircuit i/o nodes not consistent with localscope i/o nodes
SC - SubCircuit

The update() call was reinstated since it remakes the subcircuit
(and localscope) just before saving. It is important because
it makes sure that SC's i/o is consistent with localscope's i/o
(see comment) and localscope is consistent with it's referenced
actual scope.

We call rebuildCircuit() (thru update) while saving to rebuild SC's
localscope. This function is ONLY called if SC's localscope's
actual circuit's lastUpdated timestamp is after subcircuit's
last updated timestamp. Which means that circuit has changed and
SC's localscope is stale. So localscope is updated.

Another issue was that any scope's timstamp was only updating
during scheduleBackup(). schedBackup() is called at certain
time intervals and user might save circuit after making a change
before a scheduleUpdate is called and so the subcircuit will be
saved with a stale localscope.